move things around and share multi-level scrolling
